hopefully updated to work

+++ lightning_parser_lib/config_and_parser.py
@@ -192,8 +192,6 @@ def _cache_and_parse(config: LightningConfig):
     # Display available headers from the database.
     tprint("Headers:", database_parser.get_headers(config.db_path))
 
-
-
 @rf.as_remote()
 def cache_and_parse(config: LightningConfig):
     """
@@ -345,9 +343,15 @@ def delete_sql_database(config: LightningConfig):
     """
     if server_sided_config_override:
         config = server_sided_config_override
-    rf.qs.clear_hexes() # Clear hexes
-    shutil.rmtree(config.cache_dir)
-
+    time.sleep(2)
+    try:
+        if os.path.exists(config.cache_dir):
+            shutil.rmtree(config.cache_dir)
+            tprint("Deleted:", config.cache_dir)
+        else:
+            tprint("Directory does not exist:", config.cache_dir)
+    except Exception as e:
+        tprint("Error while deleting:", e)
 
 @rf.as_remote()
 def delete_pkl_cache(config: LightningConfig):
